#49-refactor-auth-with-thunk
