+++ src/store/user/selectors.ts
@@ -3,3 +3,4 @@ import { AppState } from '../types';
 export const selectUser = (state: AppState) => state.user.user;
 export const selectIsAuth = (state: AppState) => state.user.isAuth;
 export const selectIsLoading = (state: AppState) => state.user.isLoading;
+export const selectErrorMessage = (state: AppState) => state.user.errorMessage;
